Added command to execute ScrubArtifacts.pl as part of Gamegineer CruiseControl project.
Note that this change makes the CruiseControl configuration platform-dependent.  This will not execute successfully on Windows (unless Cygwin or something similar is installed).  Not really worrying about it now as we will eventually be moving off of CruiseControl and to something (Hudson/Jenkins) that hopefully has a working plugin for deleting old build artifacts.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/build/trunk@3231 897c5806-f142-0410-a293-e1bf5698dec8